Implement API changes
