+++ src/mlpack/methods/ann/CMakeLists.txt
@@ -3,10 +3,6 @@
 set(SOURCES
   ffn.hpp
   ffn_impl.hpp
-  gan.hpp
-  gan_impl.hpp
-  rbm.hpp
-  rbm_impl.hpp
   rnn.hpp
   rnn_impl.hpp
 )
