+++ src/mlpack/tests/serialization_test.cpp
@@ -34,9 +34,7 @@
 #include <mlpack/methods/lsh/lsh_search.hpp>
 #include <mlpack/methods/decision_stump/decision_stump.hpp>
 #include <mlpack/methods/lars/lars.hpp>
-#include <mlpack/methods/ann/rbm.hpp>
-#include <mlpack/methods/ann/rbm/binary_rbm_policy.hpp>
-#include <mlpack/methods/ann/rbm/spike_slab_rbm_policy.hpp>
+#include <mlpack/methods/ann/rbm/rbm.hpp>
 #include <mlpack/methods/ann/init_rules/gaussian_init.hpp>
 
 using namespace mlpack;
@@ -1706,32 +1704,31 @@ BOOST_AUTO_TEST_CASE(BinaryRBMTest)
   size_t hiddenLayerSize = 5;
   data.randu(3, 100);
 
-  BinaryRBMPolicy<> binary_rbm(data.n_rows, hiddenLayerSize);
   GaussianInitialization gaussian(0, 0.1);
-  RBM<GaussianInitialization, BinaryRBMPolicy<> > Rbm(data,
-      gaussian, binary_rbm, 1,  true, true);
-  RBM<GaussianInitialization, BinaryRBMPolicy<> > RbmXml(data,
-      gaussian, binary_rbm, 1,  true, true);
-  RBM<GaussianInitialization, BinaryRBMPolicy<> > RbmText(data,
-      gaussian, binary_rbm, 1,  true, true);
-  RBM<GaussianInitialization, BinaryRBMPolicy<> > RbmBinary(data,
-      gaussian, binary_rbm, 1,  true, true);
+  RBM<GaussianInitialization> Rbm(data, gaussian, data.n_rows, hiddenLayerSize,
+      1, 1, 1, 2, 8, 1, true);
+  RBM<GaussianInitialization> RbmXml(data, gaussian, data.n_rows,
+      hiddenLayerSize, 1, 1, 1, 2, 8, 1, true);
+  RBM<GaussianInitialization> RbmText(data, gaussian, data.n_rows,
+      hiddenLayerSize, 1, 1, 1, 2, 8, 1, true);
+  RBM<GaussianInitialization> RbmBinary(data, gaussian, data.n_rows,
+      hiddenLayerSize, 1, 1, 1, 2, 8, 1, true);
   Rbm.Reset();
 
   SerializeObjectAll(Rbm, RbmXml, RbmText, RbmBinary);
   CheckMatrices(Rbm.Parameters(), RbmXml.Parameters(), RbmText.Parameters(),
       RbmBinary.Parameters());
-  CheckMatrices(Rbm.Policy().VisibleBias(), RbmXml.Policy().VisibleBias());
-  CheckMatrices(Rbm.Policy().VisibleBias(), RbmText.Policy().VisibleBias());
-  CheckMatrices(Rbm.Policy().VisibleBias(), RbmBinary.Policy().VisibleBias());
+  CheckMatrices(Rbm.VisibleBias(), RbmXml.VisibleBias());
+  CheckMatrices(Rbm.VisibleBias(), RbmText.VisibleBias());
+  CheckMatrices(Rbm.VisibleBias(), RbmBinary.VisibleBias());
 
-  CheckMatrices(Rbm.Policy().HiddenBias(), RbmXml.Policy().HiddenBias());
-  CheckMatrices(Rbm.Policy().HiddenBias(), RbmText.Policy().HiddenBias());
-  CheckMatrices(Rbm.Policy().HiddenBias(), RbmBinary.Policy().HiddenBias());
+  CheckMatrices(Rbm.HiddenBias(), RbmXml.HiddenBias());
+  CheckMatrices(Rbm.HiddenBias(), RbmText.HiddenBias());
+  CheckMatrices(Rbm.HiddenBias(), RbmBinary.HiddenBias());
 
-  CheckMatrices(Rbm.Policy().Weight(), RbmXml.Policy().Weight());
-  CheckMatrices(Rbm.Policy().Weight(), RbmText.Policy().Weight());
-  CheckMatrices(Rbm.Policy().Weight(), RbmBinary.Policy().Weight());
+  CheckMatrices(Rbm.Weight(), RbmXml.Weight());
+  CheckMatrices(Rbm.Weight(), RbmText.Weight());
+  CheckMatrices(Rbm.Weight(), RbmBinary.Weight());
 }
 
 /**
@@ -1755,45 +1752,38 @@ BOOST_AUTO_TEST_CASE(ssRBMTest)
 
   size_t poolSize = 1;
 
-  SpikeSlabRBMPolicy<> ss_rbm(data.n_rows, hiddenLayerSize, poolSize,
-      slabPenalty, radius);
   GaussianInitialization gaussian(0, 0.1);
-  RBM<GaussianInitialization, SpikeSlabRBMPolicy<> > Rbm(data,
-      gaussian, ss_rbm, 1,  true, true);
-  RBM<GaussianInitialization, SpikeSlabRBMPolicy<>> RbmXml(data,
-      gaussian, ss_rbm, 1,  true, true);
-  RBM<GaussianInitialization, SpikeSlabRBMPolicy<>> RbmText(data,
-      gaussian, ss_rbm, 1,  true, true);
-  RBM<GaussianInitialization, SpikeSlabRBMPolicy<>> RbmBinary(data,
-      gaussian, ss_rbm, 1,  true, true);
+  RBM<GaussianInitialization, arma::mat, SpikeSlabRBM> Rbm(data, gaussian,
+      data.n_rows, hiddenLayerSize, 1, 1, 1, poolSize, slabPenalty, radius,
+      true);
+  RBM<GaussianInitialization, arma::mat, SpikeSlabRBM> RbmXml(data, gaussian,
+      data.n_rows, hiddenLayerSize, 1, 1, 1, poolSize, slabPenalty, radius,
+      true);
+  RBM<GaussianInitialization, arma::mat, SpikeSlabRBM> RbmText(data, gaussian,
+      data.n_rows, hiddenLayerSize, 1, 1, 1, poolSize, slabPenalty, radius,
+      true);
+  RBM<GaussianInitialization, arma::mat, SpikeSlabRBM> RbmBinary(data, gaussian,
+      data.n_rows, hiddenLayerSize, 1, 1, 1, poolSize, slabPenalty, radius,
+      true);
   Rbm.Reset();
-  Rbm.Policy().VisiblePenalty().fill(15);
-  Rbm.Policy().SpikeBias().ones();
+  Rbm.VisiblePenalty().fill(15);
+  Rbm.SpikeBias().ones();
 
   SerializeObjectAll(Rbm, RbmXml, RbmText, RbmBinary);
   CheckMatrices(Rbm.Parameters(), RbmXml.Parameters(), RbmText.Parameters(),
       RbmBinary.Parameters());
 
-  CheckMatrices(Rbm.Policy().VisiblePenalty(),
-      RbmXml.Policy().VisiblePenalty());
-  CheckMatrices(Rbm.Policy().VisiblePenalty(),
-      RbmText.Policy().VisiblePenalty());
-  CheckMatrices(Rbm.Policy().VisiblePenalty(),
-      RbmBinary.Policy().VisiblePenalty());
+  CheckMatrices(Rbm.VisiblePenalty(), RbmXml.VisiblePenalty());
+  CheckMatrices(Rbm.VisiblePenalty(), RbmText.VisiblePenalty());
+  CheckMatrices(Rbm.VisiblePenalty(), RbmBinary.VisiblePenalty());
 
-  CheckMatrices(Rbm.Policy().SpikeBias(),
-      RbmXml.Policy().SpikeBias());
-  CheckMatrices(Rbm.Policy().SpikeBias(),
-      RbmText.Policy().SpikeBias());
-  CheckMatrices(Rbm.Policy().SpikeBias(),
-      RbmBinary.Policy().SpikeBias());
+  CheckMatrices(Rbm.SpikeBias(), RbmXml.SpikeBias());
+  CheckMatrices(Rbm.SpikeBias(), RbmText.SpikeBias());
+  CheckMatrices(Rbm.SpikeBias(), RbmBinary.SpikeBias());
 
-  CheckMatrices(Rbm.Policy().Weight(),
-      RbmXml.Policy().Weight());
-  CheckMatrices(Rbm.Policy().Weight(),
-      RbmText.Policy().Weight());
-  CheckMatrices(Rbm.Policy().Weight(),
-      RbmBinary.Policy().Weight());
+  CheckMatrices(Rbm.WeightCube(), RbmXml.WeightCube());
+  CheckMatrices(Rbm.WeightCube(), RbmText.WeightCube());
+  CheckMatrices(Rbm.WeightCube(), RbmBinary.WeightCube());
 }
 
 BOOST_AUTO_TEST_SUITE_END();
